+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   - Example: 10.2.1.4 is the 5th version that supports khiops 10.2.1.
 - Internals: Changes in *Internals* sections are unlikely to be of interest for data scientists.
 
+## Unreleased
+
+### Added
+- (`sklearn`) `n_feature_parts` parameter to the supervised estimators 
+
 ## 11.0.0.2 - 2026-01-26
 
 ## Fixed